pyeventbt-main/pyproject.toml
[tool.poetry]
name = "pyeventbt"
version = "0.0.4"
description = "Event-driven backtesting and live trading framework for Python and MetaTrader 5"
authors = ["Marti Castany, Alain Porto"]
readme = "README.md"
license = "Apache-2.0"
homepage = "https://github.com/marticastany/pyeventbt"
repository = "https://github.com/marticastany/pyeventbt"
keywords = [
    "mt5",
    "metatrader5",
    "backtesting",
    "mt5 python",
    "event-driven",
    "trading",
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
]
packages = [{ include = "pyeventbt" }]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.1.1"
pydantic = "^2.12.3"
numpy = "^2.0.0"
polars = "^1.35.0"
pandas = "^2.2.3"
numba = "^0.62.1"
matplotlib = "^3.7.0"
scipy = "^1.10.0"
scikit-learn = "^1.3.0"
PyYAML = "^6.0"


[tool.poetry.scripts]
pyeventbt = "pyeventbt.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
